+++ RELEASES.md
@@ -1409,3 +1409,13 @@ Based on:
 - [typescript v0.7.0-alpha.9] .
 ### Releases
 - [NPM v0.7.0-alpha.9] https://www.npmjs.com/package/cribl-control-plane/v/0.7.0-alpha.9 - .
+
+## 2026-03-24 02:08:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.757.1 (2.866.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.7.0-alpha.10] .
+### Releases
+- [NPM v0.7.0-alpha.10] https://www.npmjs.com/package/cribl-control-plane/v/0.7.0-alpha.10 - .